Proper segregation of source code and tests
